create customer details step
