+++ packages/webcomponents/src/components/dispute-management/counter-dispute/counter-dispute.tsx
@@ -1,10 +1,5 @@
 import { Component, Event, EventEmitter, h, State } from "@stencil/core";
-
-export enum DisputeManagementClickEvents {
-  nextStep = 'nextStep',
-  previousStep = 'previousStep',
-  submit = 'submit',
-}
+import { DisputeManagementClickEvents } from "../dispute";
 
 @Component({
   tag: 'justifi-counter-dispute',
@@ -18,7 +13,7 @@ export class CounterDispute {
 
   componentStepMapping = [
     () => <justifi-dispute-reason ref={(el) => this.refs[0] = el}></justifi-dispute-reason>,
-    () => <justifi-upload-dispute-evidence ref={(el) => this.refs[1] = el}></justifi-upload-dispute-evidence>,
+    () => <justifi-customer-details ref={(el) => this.refs[1] = el}></justifi-customer-details>,
   ];
 
   get currentStepComponent() {
@@ -40,7 +35,7 @@ export class CounterDispute {
 
   onBack() {
     if (this.currentStep === 0) {
-      this.clickEvent.emit({ name: 'cancelDispute' });
+      this.clickEvent.emit({ name: DisputeManagementClickEvents.cancelDispute });
     } else {
       this.currentStep--;
       this.clickEvent.emit({ name: DisputeManagementClickEvents.previousStep });
@@ -62,7 +57,7 @@ export class CounterDispute {
           <div class="d-flex gap-2 mt-4 justify-content-end">
             <button class="btn btn-secondary" onClick={() => this.onBack()}>Cancel</button>
             <button class="btn btn-primary" onClick={() => this.onNext()}>
-              {this.isLastStep ? 'Next' : 'Submit Counter Dispute'}
+              {this.isLastStep ? 'Submit Counter Dispute' : 'Next'}
             </button>
           </div>
         </div>
