+++ packages/webcomponents/src/components/dispute-management/dispute-management-core.tsx
@@ -1,5 +1,6 @@
 import { Component, Event, EventEmitter, h, Listen, State } from "@stencil/core";
 import { StyledHost } from "../../ui-components";
+import { DisputeManagementClickEvents } from "./dispute";
 
 @Component({
   tag: 'justifi-dispute-management-core',
@@ -11,10 +12,10 @@ export class DisputeManagementCore {
 
   @Listen('clickEvent')
   counterDisputeHandler(event: CustomEvent) {
-    if (event.detail.name === 'counterDispute') {
+    if (event.detail.name === DisputeManagementClickEvents.counterDispute) {
       this.showCounterDispute = true;
     }
-    if (event.detail.name === 'cancelDispute') {
+    if (event.detail.name === DisputeManagementClickEvents.cancelDispute) {
       this.showCounterDispute = false;
     }
   }
