+++ packages/webcomponents/src/components/dispute-management/dispute.ts
@@ -1,0 +1,7 @@
+export enum DisputeManagementClickEvents {
+  nextStep = 'nextStep',
+  previousStep = 'previousStep',
+  cancelDispute = 'cancelDispute',
+  counterDispute = 'counterDispute',
+  submit = 'submit',
+}
